+++ pkg/api/cloud/stack/image/get.go
@@ -2,10 +2,11 @@ package image
 
 import (
 	"context"
+
 	"github.com/sitehostnz/gosh/pkg/models"
 )
 
-// Get fetches a cloud image with the spcified code. calls list and filters. basically a helper
+// Get fetches a cloud image with the spcified code. calls list and filters. basically a helper.
 func (s *Client) Get(ctx context.Context, request GetRequest) (*models.StackImage, error) {
 	listResponse, err := s.List(ctx)
 	if err != nil {