+++ pkg/api/ssh/key/update.go
@@ -3,8 +3,9 @@ package key
 import (
 	"context"
 	"fmt"
-	"github.com/sitehostnz/gosh/pkg/utils"
 	"net/url"
+
+	"github.com/sitehostnz/gosh/pkg/utils"
 )
 
 // Update an SSH Key.