+++ CHANGELOG.md
@@ -3,12 +3,16 @@ All notable changes to this project will be documented in this file. The format 
 
 ## [v0.3.x] - 2024-05-14
 ### Added
-- database list/get/update/remove for `/cloud/stack/image`
+- list/get/update/remove for `/cloud/db`
+- list/get/update/remove for `/cloud/db/user`
+- list/get/update/remove for `/cloud/db/grants`
 - cloud stack image list_all `/cloud/image`
+- add user/add/edit/update and ssh key 
 
-## [v0.3.x] - 2032-03-15
 ### Fixes
+- version updates
 - correct the ssh key update params
+- add wrappers around api return types to get bools, not ints.
 
 ## [v0.3.4] - 2024-03-12
 ### Added